FIX | Preserve visual newlines in extended full text
Previously, newlines in the full text were not respected visually. While present in the html, the full text didn't show newlines because of the default value of the css "white-space" rule. This RB gives ful text the "white-space: pre-line;" rule, which preserves newlines.

The behavior of stripText, which should apply only to the clamped version of the text, is unaffected.